sdk version in demo scene, improved transition panel, react template is full screen
