--- Demo/sequence-demo/StreamingAssets/UnityServicesProjectConfiguration.json
@@ -1,1 +1,0 @@
-{"Keys":["com.unity.services.core.version"],"Values":[{"m_Value":"1.4.0","m_IsReadOnly":true}]}
