--- Demo/sequence-demo/src/App.css
@@ -1,65 +1,0 @@
-.outer-container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100vw;
-  height: 100vh;
-}
-
-.container {
-  position: relative;
-  /* The container determains the size. */
-  height: 90%;
-  max-width: 100%;
-  aspect-ratio: 16 / 9;
-}
-
-.container > .loading-overlay {
-  /* We'll render the overlay on top of the Unity Application. */
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: grey;
-  /* We'll set the following Flex properties in order to center the text. */
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.container > .unity {
-  /* The Unity Application matches it size to the container. */
-  width: 100%;
-  height: 100%;
-}
-
-.login-outer-container {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: rgba(0, 0, 0, 0.9);
-}
-
-.login-container {
-  position: relative;
-  width: 300px;
-  height: 200px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  border-radius: 6px;
-  background-color: white;
-}
-
-.login-title {
-  display: block;
-  color: black;
-  font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
-  margin-bottom: 40px;
-}
