--- Packages/Sequence-Unity/Sequence/Samples/Setup/Resources/SequenceConfig.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 111c4b287f2d64d968c331b0c34de28e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
