+++ Packages/Sequence-Unity/Sequence/SequenceFrontend/Prefabs/DemoButton.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2b585c6ea5ce547c3b5638accfd28750
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
