+++ Packages/Sequence-Unity/Sequence/SequenceFrontend/Scripts/UI/EmailSignedInText.cs
@@ -1,7 +1,5 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using Sequence.Authentication;
+using Sequence.Config;
 using Sequence.EmbeddedWallet;
 using TMPro;
 using UnityEngine;
@@ -10,11 +8,18 @@ namespace Sequence.Demo
 {
     public class EmailSignedInText : MonoBehaviour
     {
+        private TextMeshProUGUI _text;
         private SequenceWallet _wallet;
 
-        void Awake()
+        private void Awake()
         {
             SequenceWallet.OnWalletCreated += OnWalletCreated;
+        }
+
+        private void Start()
+        {
+            _text = GetComponent<TextMeshProUGUI>();
+            OnWalletCreated(null);
         }
 
         private void OnDestroy()
@@ -24,14 +29,14 @@ namespace Sequence.Demo
 
         private void OnWalletCreated(SequenceWallet wallet)
         {
-            TextMeshProUGUI text = GetComponent<TextMeshProUGUI>();
-            text.text = "Logged in as: " + PlayerPrefs.GetString(OpenIdAuthenticator.LoginEmail);
             _wallet = wallet;
-        }
-
-        public void SignOut()
-        {
-            _wallet.DropThisSession();
+            var sdkVersion = SequenceConfig.GetConfig().WaaSVersion;
+            var email = PlayerPrefs.GetString(OpenIdAuthenticator.LoginEmail);
+            
+            if (_wallet == null || string.IsNullOrEmpty(email))
+                _text.text = sdkVersion;
+            else
+                _text.text = $"Logged in as: {email}, {sdkVersion}";
         }
     }
 }
